Refactor background service imports and remove unused getTextFromDeepseek and Prompt files
