+++ extension/entrypoints/background/Prompt.ts
